requirements.txt
streamlit
langchain
langchain-community
langchain-text-splitters
langchain-google-genai
pypdf
faiss-cpu
python-dotenv
tiktoken